+++ website/app/_meta.global.ts
@@ -1,3 +1,26 @@
+import { rules } from '@graphql-eslint/eslint-plugin';
+
+const { schemaRules, operationsRules, schemaAndOperationsRules } = Object.entries(rules)
+  .sort(([a], [b]) => a.localeCompare(b))
+  .reduce<{
+    schemaRules: string[];
+    operationsRules: string[];
+    schemaAndOperationsRules: [];
+  }>(
+    (acc, [ruleId, curr]) => {
+      const { category } = curr.meta.docs;
+      if (category === 'Schema') {
+        acc.schemaRules.push(ruleId);
+      } else if (category === 'Operations') {
+        acc.operationsRules.push(ruleId);
+      } else {
+        acc.schemaAndOperationsRules.push(ruleId);
+      }
+      return acc;
+    },
+    { schemaRules: [], operationsRules: [], schemaAndOperationsRules: [] },
+  );
+
 export default {
   docs: {
     title: 'Documentation',
@@ -26,76 +49,28 @@ export default {
     title: 'Rules',
     type: 'page',
     items: {
-      index: { title: 'Overview', theme: { layout: 'full' } },
-      prettier: '`prettier` Rule',
-      'deprecated-rules': 'Deprecated Rules',
-      '---1': { title: 'Schema & Operations Rules', type: 'separator' },
-      alphabetize: '',
-      'known-argument-names': '',
-      'known-directives': '',
-      'known-type-names': '',
-      'naming-convention': '',
-      'provided-required-arguments': '',
-      'unique-directive-names-per-location': '',
-      '---2': { title: 'Schema Rules', type: 'separator' },
-      'description-style': '',
-      'input-name': '',
-      'lone-schema-definition': '',
-      'no-hashtag-description': '',
-      'no-root-type': '',
-      'no-scalar-result-type-on-mutation': '',
-      'no-typename-prefix': '',
-      'no-unreachable-types': '',
-      'no-unused-fields': '',
-      'possible-type-extension': '',
-      'relay-arguments': '',
-      'relay-connection-types': '',
-      'relay-edge-types': '',
-      'relay-page-info': '',
-      'require-deprecation-date': '',
-      'require-deprecation-reason': '',
-      'require-description': '',
-      'require-field-of-type-query-in-mutation-result': '',
-      'require-nullable-fields-with-oneof': '',
-      'require-nullable-result-in-root': '',
-      'require-type-pattern-with-oneof': '',
-      'strict-id-in-types': '',
-      'unique-directive-names': '',
-      'unique-enum-value-names': '',
-      'unique-field-definition-names': '',
-      'unique-operation-types': '',
-      'unique-type-names': '',
-      '---3': { title: 'Operations Rules', type: 'separator' },
-      'executable-definitions': '',
-      'fields-on-correct-type': '',
-      'fragments-on-composite-type': '',
-      'known-fragment-names': '',
-      'lone-anonymous-operation': '',
-      'lone-executable-definition': '',
-      'match-document-filename': '',
-      'no-anonymous-operations': '',
-      'no-deprecated': '',
-      'no-duplicate-fields': '',
-      'no-fragment-cycles': '',
-      'no-one-place-fragments': '',
-      'no-undefined-variables': '',
-      'no-unused-fragments': '',
-      'no-unused-variables': '',
-      'one-field-subscriptions': '',
-      'overlapping-fields-can-be-merged': '',
-      'possible-fragment-spread': '',
-      'require-import-fragment': '',
-      'require-selections': '',
-      'scalar-leafs': '',
-      'selection-set-depth': '',
-      'unique-argument-names': '',
-      'unique-fragment-name': '',
-      'unique-input-field-names': '',
-      'unique-operation-name': '',
-      'unique-variable-names': '',
-      'value-literals-of-correct-type': '',
-      'variables-are-input-types': '',
-      'variables-in-allowed-position': '',
+      index: {
+        theme: {
+          layout: 'full',
+        },
+      },
+      prettier: '',
+      'deprecated-rules': '',
+      _1: {
+        title: 'Schema & Operations Rules',
+        type: 'separator',
+      },
+      ...Object.fromEntries(schemaAndOperationsRules.map(key => [key, ''])),
+      _2: {
+        title: 'Schema Rules',
+        type: 'separator',
+      },
+      ...Object.fromEntries(schemaRules.map(key => [key, ''])),
+      _3: {
+        title: 'Operations Rules',
+        type: 'separator',
+      },
+      ...Object.fromEntries(operationsRules.map(key => [key, ''])),
     },
   },
   play: {
